Edit in Userstyles with the Stylish editor
Add CodeMirror to `textarea.css` in https://userstyles.org/styles/*/edit
pages. This is a trial balloon. Before it's pulled the following changes
should be considered:
(1) The editor functionality should be removed from `edit.js` into its own
file for use with both `edit.js` and `install.js`. What you have here is
the "old" editor without the recent improvements.
(2) Editing in Userstyles with CodeMirror should be optional. It could be
associated with a UI control, but really it's there as a circuit breaker
to preserve site functionality in case of a CodeMirror catastrophe.
(3) If (2), a preference value must be passed to the content script before
the editor instance can start.
(4) The editor controls are part of `edit.html`. If they're wanted on the
Userstyles editor page, they'll need to be copied or abstracted from it.